+++ {{cookiecutter.project_name}}/{{cookiecutter.package_name}}/utils.py
@@ -1,17 +1,22 @@
+"""Helper functions for the package.
+"""
 import importlib.metadata
 from typing import Union
 
 
 def get_version_tuple() -> tuple:
-    def as_integer(x: str) -> Union[int, str]:
+    """Get the package version as a tuple.
+
+    Returns:
+        tuple: the package version as a tuple following semantic versioning.
+    """
+
+    def as_integer(x: str) -> Union[int, str]:  # pylint: disable=C0103
         try:
             return int(x)
         except ValueError:
             return x
 
     return tuple(
-        as_integer(v)
-        for v in importlib.metadata.version("{{ cookiecutter.package_name }}")
-        .strip()
-        .split(".")
+        as_integer(v) for v in importlib.metadata.version("bw2ui").strip().split(".")
     )